+++ CHANGELOG.md
@@ -10,9 +10,13 @@
       (fixes #205)
     * Engine and Rule events `on('success')`, `on('failure')`, and Rule callbacks `onSuccess` and `onFailure` now honor returned promises; any event handler that returns a promise will be waited upon to resolve before engine execution continues. (fixes #235)
     * Private `rule.event` property renamed. Use `rule.getEvent()` to avoid breaking changes in the future.
-    * The 'success-events' fact used to store successful events has been converted to an internal data structure and will no longer appear in the almanac's facts. (fixes #187)
+    * The `success-events` fact used to store successful events has been converted to an internal data structure and will no longer appear in the almanac's facts. (fixes #187)
   * NEW FEATURES
     * Engine constructor now accepts a `pathResolver` option for resolving condition `path` properties. Read more [here](./docs/rules.md#condition-helpers-custom-path-resolver). (fixes #210)
+    * Engine.run() now returns three additional data structures:
+      * `failureEvents`, an array of all failed rules events. (fixes #192)
+      * `results`, an array of RuleResults for each successful rule (fixes #216)
+      * `failureResults`, an array of RuleResults for each failed rule
 
 
 #### 5.3.0 / 2020-12-02